Time: 4 ms (16.19%), Space: 7.7 MB (57.35%) - LeetHub

+++ 0028-find-the-index-of-the-first-occurrence-in-a-string/0028-find-the-index-of-the-first-occurrence-in-a-string.cpp
@@ -1,8 +1,8 @@
 class Solution {
 public:
     int strStr(string haystack, string needle) {
-        int pos = haystack.find(needle);
-        if (pos != string::npos) { /*(find) if it found the strin, will return first index that string start at else it will return (string::npos)*/
+        size_t pos = haystack.find(needle);
+        if (pos != string::npos) { /*(find) if it found the string, will return first index that string start at else it will return (string::npos)*/
             return pos;
         } else {
             return -1;
